add favorites display on textarea
